+++ neo/sys/win32/win_net.cpp
@@ -404,8 +404,8 @@ void NET_OpenSocks( int port ) {
 		int		plen;
 
 		// build the request
-		ulen = strlen( net_socksUsername.GetString() );
-		plen = strlen( net_socksPassword.GetString() );
+		ulen = (int)strlen( net_socksUsername.GetString() );
+		plen = (int)strlen( net_socksPassword.GetString() );
 
 		buf[0] = 1;		// username/password authentication version
 		buf[1] = ulen;
